refactor(core): relocate glob cluster to julie_core::glob (Phase 1 T1)
Move the 5-symbol glob cluster (matches_glob_pattern, PatternMatcher,
CompiledPatterns, split_top_level_commas, compile_patterns) from
src/tools/search/query.rs into a new crates/julie-core/src/glob.rs.

Re-export shim in query.rs keeps all 30 callers compiling unchanged
(crate::tools::search::query::matches_glob_pattern still resolves).
Adds globset = "0.4.16" to julie-core's dependencies.

Verified:
- cargo check: clean workspace-wide
- cargo nextest run -p julie-core: 125/125 (incl. 5 new glob unit tests)
- cargo nextest run --lib test_glob_pattern: 7/7 integration tests via shim
