+++ .memories/2026-06-03/222114_2f44.md
@@ -1,0 +1,77 @@
+---
+id: checkpoint_2f44ca3f
+timestamp: 2026-06-03T22:21:14.947Z
+tags:
+  - julie-rescue
+  - phase1
+  - T1
+  - glob
+  - refactor
+  - julie-core
+git:
+  branch: julie-rescue
+  commit: 6f859383
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - Cargo.lock
+    - crates/julie-core/Cargo.toml
+    - crates/julie-core/src/glob.rs
+    - crates/julie-core/src/lib.rs
+    - src/tools/search/query.rs
+summary: "T1 Complete: Glob cluster relocated to `julie_core::glob`"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo check: clean (48.92s)"
+  - "cargo nextest run -p julie-core: 125/125 passed"
+  - "cargo nextest run --lib test_glob_pattern: 7/7 passed"
+symbols:
+  - matches_glob_pattern
+  - PatternMatcher
+  - CompiledPatterns
+  - split_top_level_commas
+  - compile_patterns
+  - julie_core::glob
+next: "T2: Scaffold julie-index crate, move search+analysis+languages modules in"
+---
+
+## T1 Complete: Glob cluster relocated to `julie_core::glob`
+
+### WHAT
+Moved the 5-symbol glob cluster from `src/tools/search/query.rs` into a new `crates/julie-core/src/glob.rs` leaf module.
+
+**Symbols moved:**
+- `matches_glob_pattern` (pub fn)
+- `PatternMatcher` (private enum + impl)
+- `CompiledPatterns` (private struct + impl)
+- `split_top_level_commas` (private fn)
+- `compile_patterns` (private fn)
+
+### WHY
+Phase 1 of Julie Rescue: extracting a `julie-index` crate requires the glob helper to live below both `tools` and the future `search`/`analysis` modules. Moving it down to the `julie-core` leaf makes it a shared dependency.
+
+### HOW
+- Created `crates/julie-core/src/glob.rs` with verbatim code (only import adjustments)
+- Added `globset = "0.4.16"` to `crates/julie-core/Cargo.toml`
+- Added `pub mod glob;` to `crates/julie-core/src/lib.rs`
+- Removed 5 symbols from `src/tools/search/query.rs`; dropped `use globset` and `use tracing::warn` imports (no longer needed there)
+- Added `pub use julie_core::glob::matches_glob_pattern;` re-export shim so all 30 callers continue to resolve via `crate::tools::search::query::matches_glob_pattern` unchanged
+
+### IMPACT
+- Zero call-site edits needed — re-export shim keeps the old path working
+- MCP orientation confirmed: all 4 private symbols had zero external refs; `matches_glob_pattern` had 30 refs all routing through the shim
+- 5 new unit tests added in `crates/julie-core/src/glob.rs`
+
+### Evidence
+- `cargo check` — clean workspace-wide
+- `cargo nextest run -p julie-core` — 125/125 passed (incl. 5 new glob tests)
+- `cargo nextest run --lib test_glob_pattern` — 7/7 integration glob tests passed via shim
